Creating find user
Creating a new hook to understand better the getDocs hook, this includes failed code attempts, I think I have now figured it out and will polish this.

my mistake: I tried to in the hook save the result from the getDocs into a useState but I had to simply save it as a const and return the const.