Update to AnimeDescription & useFirestore
Added new function to useFirestore to update description rather than comment.
I need to update the comment labelling as it is too generalized and needs to be specific to updating the comment.

NEXT: I need to show the description at the top and give the user the option to read all potential descriptions of the anime and create a ranking system for the future.